修复FormulaPP的异常

+++ src/memect/pdf/formula_pp.py
@@ -240,7 +240,7 @@ class _UniMERNetDecode:
         return self._normalize(text)
 
     def __call__(self, preds: np.ndarray) -> list[str]:
-        token_ids = np.array(preds)
+        token_ids = np.array(preds).astype(np.int32)
         results = []
         for tok_id in token_ids:
             end_idx = np.argwhere(tok_id == self._eos_token_id)
